requirements.txt
fastapi==0.115.6
SQLAlchemy==2.0.36
pydantic==2.10.4
passlib==1.7.4
python-jose==3.3.0